Renamed SLDA to ShrinkageLDA and moved it to lda.py.
